login auth updates
I've added debugging logs to both components. Now try logging in again and check the browser console. We should see:
The token being stored
The redirect being attempted
The ProtectedRoute component's evaluation of the token
This will help us understand:
If the token is being stored correctly
If the redirect is happening
Why the ProtectedRoute might be rejecting the token